+++ src/main/java/net/yigitguven/profile/ProfileCommand.java
@@ -1,0 +1,24 @@
+package net.yigitguven.profile;
+
+import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.arguments.StringArgumentType;
+import net.yigitguven.profile.network.ProfileNetwork;
+import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.commands.Commands;
+import net.minecraft.commands.arguments.EntityArgument;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
+
+public class ProfileCommand {
+    public static void register(CommandDispatcher<CommandSourceStack> dispatcher) {
+        dispatcher.register(Commands.literal("profile")
+                .then(Commands.literal("view")
+                        .then(Commands.argument("player", EntityArgument.player())
+                                .executes(context -> {
+                                    ServerPlayer sender = context.getSource().getPlayerOrException();
+                                    ServerPlayer target = EntityArgument.getPlayer(context, "player");
+                                    ProfileNetwork.sendProfileToClient(sender, target);
+                                    return 1;
+                                }))));
+    }
+}